+++ easy3d/algo/extrusion.cpp
@@ -26,7 +26,7 @@
 #include <easy3d/algo/extrusion.h>
 
 #include <easy3d/core/surface_mesh.h>
-#include <easy3d/core/manifold_builder.h>
+#include <easy3d/core/surface_mesh_builder.h>
 #include <easy3d/algo/tessellator.h>
 #include <easy3d/util/logging.h>
 
@@ -115,8 +115,8 @@ namespace easy3d {
             // the vertex index starts from 0 for each character.
             const int offset = mesh->n_vertices();
 
-            // use ManifoldBuilder (ensuring the final model is manifold).
-            ManifoldBuilder builder(mesh);
+            // use SurfaceMeshBuilder (ensuring the final model is manifold).
+            SurfaceMeshBuilder builder(mesh);
             builder.begin_surface();
 
             const auto &final_vertices = tessellator.vertices();