+++ easy3d/algo/surface_mesh_polygonization.cpp
@@ -24,7 +24,7 @@
 
 #include <easy3d/algo/surface_mesh_polygonization.h>
 #include <easy3d/algo/surface_mesh_enumerator.h>
-#include <easy3d/core/manifold_builder.h>
+#include <easy3d/core/surface_mesh_builder.h>
 #include <easy3d/util/logging.h>
 
 
@@ -93,7 +93,7 @@ namespace easy3d {
         }
 
         mesh->clear();
-        ManifoldBuilder builder(mesh);
+        SurfaceMeshBuilder builder(mesh);
         builder.begin_surface();
 
         for (auto v : model.vertices())