+++ easy3d/algo_ext/overlapping_faces.cpp
@@ -28,7 +28,7 @@
 #include <unordered_map>
 
 #include <easy3d/core/surface_mesh.h>
-#include <easy3d/core/manifold_builder.h>
+#include <easy3d/core/surface_mesh_builder.h>
 #include <easy3d/util/logging.h>
 
 
@@ -241,7 +241,7 @@ namespace easy3d {
         input->clear();
         auto points = mesh.get_vertex_property<vec3>("v:point");
 
-        ManifoldBuilder builder(input);
+        SurfaceMeshBuilder builder(input);
         builder.begin_surface();
 
         for (auto v : mesh.vertices())