+++ easy3d/fileio/surface_mesh_io_stl.cpp
@@ -31,7 +31,7 @@
 #include <cfloat>
 
 #include <easy3d/core/surface_mesh.h>
-#include <easy3d/core/manifold_builder.h>
+#include <easy3d/core/surface_mesh_builder.h>
 #include <easy3d/util/logging.h>
 
 
@@ -103,7 +103,7 @@ namespace easy3d {
 			// clear mesh
 			mesh->clear();
 
-            ManifoldBuilder builder(mesh);
+            SurfaceMeshBuilder builder(mesh);
             builder.begin_surface();
 
 			// open file (in ASCII mode)