+++ ChangeLog.md
@@ -9,7 +9,7 @@
   - reorganize the tutorials (because I am planning to add more tutorials which needs to be better grouped).
   
 #### 27/01/2020
-  - now loading models from files is done through ManifoldBuilder that resolves non-manifoldness.
+  - now loading models from files is done through SurfaceMeshBuilder that resolves non-manifoldness.
 
 #### 20/12/2019
   - now all image io is managed by ImageIO.